refactor(telemetry): inline AI feature config into the telemetry job
Drop the extra helper file and telemetry-only tests; resolve config through the existing product helpers in index.ts.
